Set the default value for URL schemes in form inputs.
When a URL field is filled out with a URL that doesn't include a scheme,
Django adds one. They're going to be changing the default of this from
http to https in Django 6.0, and in the meantime, every time it gets
instantiated, it spews a warning.

This change adds the transitional setting to opt-in to the new behavior
and suppress the warning.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/14815/