Only include "validation-test" search backend when actually validating.
While debugging something with search, I noticed that during indexing,
each model was getting indexed twice. It turns out this is because of
the "validation-test" key we had inside the search config, even
though its actual configuration was empty.

This change makes it so we only define that backend for the duration of
actually validating, rather than always having it in the config.

Testing Done:
Ran an index and saw that it only indexed each model once.

Reviewed at https://reviews.reviewboard.org/r/14834/